backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
openpyxl==3.1.5
xlrd==1.2.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0

